made blinding test labels with full annotations optional based on their presence

+++ main.py
@@ -47,17 +47,19 @@ def main():
         print "main: ... done"
 
         # Create a new labels.pickle that erases the labels of the active training set for test purposes.
-        print "main: creating new labels.pickle that blinds the active training set for this test..."
-        with open(os.path.join(kb_perception_source_dir, 'full_annotations.pickle'), 'rb') as f:
-            fa = pickle.load(f)
-        with open(os.path.join(kb_perception_source_dir, 'labels.pickle'), 'wb') as f:
-            labels = []
-            for oidx in fa:
-                if oidx not in active_train_set:
-                    for pidx in range(len(fa[oidx])):
-                        labels.append((pidx, oidx, fa[oidx][pidx]))
-            pickle.dump(labels, f)
-        print "main: ... done"
+        full_annotation_fn = os.path.join(kb_perception_source_dir, 'full_annotations.pickle')
+        if os.path.isfile(full_annotation_fn):
+            print "main: creating new labels.pickle that blinds the active training set for this test..."
+            with open(full_annotation_fn, 'rb') as f:
+                fa = pickle.load(f)
+            with open(os.path.join(kb_perception_source_dir, 'labels.pickle'), 'wb') as f:
+                labels = []
+                for oidx in fa:
+                    if oidx not in active_train_set:
+                        for pidx in range(len(fa[oidx])):
+                            labels.append((pidx, oidx, fa[oidx][pidx]))
+                pickle.dump(labels, f)
+            print "main: ... done"
 
         # Instantiate a grounder.
         print "main: instantiating grounder..."
